Updates to main page and JS file.
Working on getting AJAX call set-up and working for the form "submit".
Changed the compare button to a normal button and no longer a submit.
Removed the form method. Updated the JS file to add a click event
handler to the compare button which triggers an AJAX request to the API
page passing along the input values.